+++ crates/astria-sequencer-relayer/CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix memory leak in metrics [2221](https://github.com/astriaorg/astria/pull/2221).
+
 ## [1.0.1] - 2025-03-06
 
 ### Changed