+++ crates/astria-sequencer/CHANGELOG.md
@@ -20,6 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Disallow ICS20 transfers using denoms which are not allowed fee assets post-Blackburn
   upgrade [#2228](https://github.com/astriaorg/astria/pull/2228).
 
+### Fixed
+
+- Fix memory leak in metrics [2221](https://github.com/astriaorg/astria/pull/2221).
+
 ## [3.0.0] - 2025-05-21
 
 ### Added